requirements.txt
pyyaml
colorama
questionary
pynetbox